+++ validator/src/cli.rs
@@ -332,7 +332,11 @@ pub fn app<'a>(version: &'a str, default_args: &'a DefaultArgs) -> App<'a, 'a> {
                 .long("snapshots")
                 .value_name("DIR")
                 .takes_value(true)
-                .help("Use DIR as snapshot location [default: <LEDGER>/snapshots]"),
+                .help(
+                    "Use DIR as the base location for snapshots. \
+                     A subdirectory named \"snapshots\" will be created. \
+                     [default: --ledger value]",
+                 ),
         )
         .arg(
             Arg::with_name(use_snapshot_archives_at_startup::cli::NAME)
@@ -344,13 +348,23 @@ pub fn app<'a>(version: &'a str, default_args: &'a DefaultArgs) -> App<'a, 'a> {
                 .long_help(use_snapshot_archives_at_startup::cli::LONG_HELP),
         )
         .arg(
+            Arg::with_name("full_snapshot_archive_path")
+                .long("full-snapshot-archive-path")
+                .value_name("DIR")
+                .takes_value(true)
+                .help(
+                    "Use DIR as full snapshot archives location \
+                     [default: --snapshots value]",
+                 ),
+        )
+        .arg(
             Arg::with_name("incremental_snapshot_archive_path")
                 .long("incremental-snapshot-archive-path")
                 .conflicts_with("no-incremental-snapshots")
                 .value_name("DIR")
                 .takes_value(true)
                 .help(
-                    "Use DIR as separate location for incremental snapshot archives \
+                    "Use DIR as incremental snapshot archives location \
                      [default: --snapshots value]",
                 ),
         )